+++ phone_server/user_list.go
@@ -47,20 +47,23 @@ func (users *UserList) remove(usr *User) {
 
 // Sends a state data update to all users
 func (users *UserList) updateData() {
-    // TODO: calculate coordinates in client side space
     asteroidMap.copyC <- nil
     playerShip.getC <- nil
     asteroidData := <-asteroidMap.copyC
     playerShipData := <-playerShip.getC
+    // Transform asteroid coordinates into phone screen space
     for _, asteroid := range asteroidData {
+        // Centre grid around player ship
         asteroid.posX -= playerShipData.posX
         asteroid.posY -= playerShipData.posY
-        newX := float64(asteroid.posX)*math.Cos((180+playerShipData.rot)*(math.Pi/180)) - float64(asteroid.posY)*math.Sin((180+playerShipData.rot)*(math.Pi/180))
-        newY := float64(asteroid.posX)*math.Sin((180+playerShipData.rot)*(math.Pi/180)) + float64(asteroid.posY)*math.Cos((180+playerShipData.rot)*(math.Pi/180))
-        //fmt.Println("X: ",asteroid.posX, " New X: ", newX, " Y: ", asteroid.posY, " New Y: ", newY)
-        asteroid.posX = int(newX) + 50
-        asteroid.posY = int(newY) + 50
+        // Rotate grid so ship is pointing north on the screen
+        newX := asteroid.posX*math.Cos((180+playerShipData.rot)*(math.Pi/180)) - asteroid.posY*math.Sin((180+playerShipData.rot)*(math.Pi/180))
+        newY := asteroid.posX*math.Sin((180+playerShipData.rot)*(math.Pi/180)) + asteroid.posY*math.Cos((180+playerShipData.rot)*(math.Pi/180))
+        // Translate grid so ship is in the centre
+        asteroid.posX = newX + 50
+        asteroid.posY = newY + 50
     }
+    // Send updated data
     for _, usr := range users.l {
         usr.sendDataUpdate(asteroidData)
     }
